Added colr tool, enhanced gradient, strip_codes,..
This module can now be used as a tool:
    * Docopt is required to use the tool, but not the library itself.
    `python3 -m colr "My Text" -f "green" -b "white" -s "bright"`

For more options:
    `python3 -m colr ---help`

`strip_codes()` was added as a convenience function, and is being used
for stripping codes when calling `gradient()` on `self.data`.

`gradient()` can now be called with no arguments. It will operate on
`self.data` when no text is given, and join `self.data` with the
new gradient when text is given.

`auto_disable()`, `enable()`, and `disable()` just use a boolean flag.
No need to monkey-patch (or re-patch) anything when using them.
Just calling `auto_disable()` should be enough.

`_str_just()` was cleaned a little.

The test runs in `test_colr.py` have a few more sanity checks,
mostly dealing with these new changes. Unit tests are coming.