Use Sequence base, and add char/code navigation.
`ChainedBase` now inherits from `Sequence` instead of `object`.
The change was made to be more explicit about what kind of object
the `ChainedBase` is, and to better satisfy type checks.

Helper functions were added in `colr/base.py` to help navigate
strings of chars and escaope-codes. They are currently not used
in `Colr` itself, but may be in the future.